Includes a section for use on player entities
A minor tweak is necessary in order to use this method on player entities. This update adds an example of said tweak. It could be useful for people who are following this tutorial for the purpose of tracking player deaths, as the original does not take into account the requirement to reset the state change.