Updated, re-formatted.
I tried putting the buttons in a table ( symbol | name ), but turns out I don't really know the names of most of these. Thus I figured users would be able to map the symbols to the now embedded images themselves. It might change with an external contribution in the future. For now, it's usable.