Changed the method but updated the tutorial
Thank fond for pointing this out, I am not sure i got the mark down correct, but this process does work i tested it. it unfortunately doesnt seem to work for physical texture sets, the physical texture do not get loaded. i posted the bug in here.

I am not the best at mark down, so if there are errors i apologize