Added minecraft:wither_skull_dangerous identifier
With some messing around, I realized that minecraft:wither_skull_dangerous is a runtime identifier, with all of the effects I noted in the description for it. There may be more entities zombies that are affected weirdly by being killed by a projectile with it, but I haven't found any(only tested some passive mobs, skeletons, and husks).